QoL Improvements to tools
QoL improvements to configuring the mdate variable. The {:02d} takes an integer input and pads it to a 2 character string with zeros in front of it if needed.

Also, changed the return for spt.get_init_time(model) to condense it to one output rather than calling the function 2 separate times in a row.